requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
geographiclib==2.0
geopy==2.4.1
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
wfastcgi==3.0.0
